Turn class-redirector into a Task
