+++ tools/class-redirector/class-redirector/class-redirector.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
+    <OutputType>Library</OutputType>
     <TargetFramework>net7.0</TargetFramework>
     <RootNamespace>class_redirector</RootNamespace>
     <ImplicitUsings>enable</ImplicitUsings>
@@ -11,6 +11,7 @@
   <ItemGroup>
     <PackageReference Include="Mono.Cecil" Version="0.11.4" />
     <PackageReference Include="Mono.Options" Version="6.12.0.148" />
+    <PackageReference Include="Microsoft.Build.Utilities.Core" Version="17.5.0" />
   </ItemGroup>
   <ItemGroup>
     <Compile Include="..\..\common\CSToObjCMap.cs">
